+++ applications/external/apple_ble_spam/lib/continuity/continuity.h
@@ -44,6 +44,6 @@ typedef struct {
 
 const char* continuity_get_type_name(ContinuityType type);
 
-size_t continuity_get_packet_size(ContinuityType type);
+uint8_t continuity_get_packet_size(ContinuityType type);
 
 void continuity_generate_packet(const ContinuityMsg* msg, uint8_t* packet);
